[SMK-31] / EmerLight-GPIO Move Location
GPIO.c, GPIO.h -> IO 폴더로 이동
emer_light.c, emer_light.h -> IO 폴더로 이동
GPIO.c 필요없는 내용(스위치 관련) 삭제
