+++ platforms/tc375/BSW/Driver/GPIO.h
@@ -1,5 +1,5 @@
-#ifndef BSW_IO_GPIO_H_
-#define BSW_IO_GPIO_H_
+#ifndef BSW_DRIVER_GPIO_H_
+#define BSW_DRIVER_GPIO_H_
 
 /* Init functions */
 void GPIO_Init(void);
@@ -11,4 +11,4 @@ int GPIO_getSW2(void);
 int GPIO_getSW3(void);
 int GPIO_getSWxDebounce(int sw_num);
 
-#endif /* BSW_IO_GPIO_H_ */
+#endif /* BSW_DRIVER_GPIO_H_ */
