+++ platforms/tc375/BSW/IO/emer_light.c
@@ -1,5 +1,5 @@
-#include "emer_light.h"
-#include "GPIO.h"
+#include <emer_light.h>
+#include <GPIO.h>
 #include "stm.h"
 
 void Emer_Light_Init(void)
