+++ platforms/tc375/system_init.h
@@ -1,6 +1,8 @@
 #ifndef SYSTEM_INIT_H_
 #define SYSTEM_INIT_H_
 
+#include <emer_light.h>
+#include <GPIO.h>
 #include "Ifx_Types.h"
 #include "IfxCpu.h"
 #include "IfxScuWdt.h"
@@ -11,8 +13,6 @@
 #include "gpt12.h"
 #include "can.h"
 #include "asclin.h"
-#include "GPIO.h"
-#include "Emer_Light.h"
 #include "stm.h"
 
 void System_Init (void);
